# Parse() performance optimization