@marselester 2024-12-04
> the parallelization we have, which you are removing, looks useful if the parsing is slow

Oh, you're right! Maybe checking the cache first, and then running a parser in a goroutine on a cache miss would be beneficial.